Using tpolecat compiler flags plugin, field for enumeratum and general tidy
