+++ app/uk/gov/hmrc/disguisedremunerationfrontend/aa_data/LoanDetails.scala
@@ -20,14 +20,11 @@ import java.time.LocalDate
 
 import ltbs.uniform._
 import play.api.i18n.Messages
-import uk.gov.hmrc.disguisedremunerationfrontend.controllers.{YesNoUnknown, YesNoUnknownWrittenOff}
 import uk.gov.hmrc.disguisedremunerationfrontend.data.Scheme.moneyRegex
 
 import scala.language.higherKinds
 import cats.data.NonEmptyList
 import cats.implicits._
-import ltbs.uniform.{::, Language, NilTypes, _}
-import play.api.i18n.{Messages => _}
 
 case class LoanDetails(
     year: Int,
@@ -129,7 +126,7 @@ object LoanDetails {
         default = default.map(x => x.isWrittenOff match {
           case Some(x: List[String]) if x == List(YesNoUnknown.CUnknown.entryName) => YesNoUnknownWrittenOff.Unknown
           case Some(x: List[String]) if x == List(YesNoUnknown.BNo.entryName) => YesNoUnknownWrittenOff.No
-          case Some(x: List[String]) => YesNoUnknownWrittenOff(x.some)
+          case Some(x: List[String]) => YesNoUnknownWrittenOff(x)
         }),
         customContent = Map(
           "written-off.heading" ->
