+++ project/plugins.sbt
@@ -19,3 +19,5 @@ addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.6.20")
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.5.0")
 
 addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "1.0.0")
+
+addSbtPlugin("io.github.davidgregory084" % "sbt-tpolecat" % "0.1.8")
